Remove appending class proprties to property cache.  Correct element instance logic in get_path
Element class __getattr__ method appends data to class property cache, however those properties aren't always valid, such as if a single instance of a key is later converted to a list. This results in a mangled tree where the data is still in the tree, but attempts to access it now reference the wrong objects.

In get_path, there is a check while descending the Element class tree where a check is performed to see if a matching element is instantiated or not, however, a __bool__ override in the Element class returns False when in the case of get_path, the result should be true. Changing the logic to not use the boolean from the class, and instead just check if the new path has been stored or is still the default value of None.