Fix Element class properties, Fix add method to include parent, Update representation for clarity
Element class properties are referencing a static type, resulting in all instances of the class being updated if the static type is replaced.  Updated the Element representation method to use the element name property if the cdata is empty to make diagnostics easier.  This fixes the case where the representation looks like [, , ,].